# web/openapi/wc26.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WC26 Transfer Intelligence API",
    "summary": "Football recruitment intelligence powered by World Cup data.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/players/search": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Search the player catalogue",
        "description": "Return lightweight players matching the supplied name query.",
        "operationId": "search_player_catalogue_api_v1_players_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 100,
              "description": "Case- and diacritic-insensitive player-name query.",
              "examples": [
                "olise"
              ],
              "title": "Q"
            },
            "description": "Case- and diacritic-insensitive player-name query."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "minimum": 1,
              "description": "Maximum number of matching players.",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of matching players."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "The player-search parameters are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The player profile datasets are missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Player search failed unexpectedly.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/players/{player_id}": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get one player profile",
        "description": "Return the detailed profile for one player ID.",
        "operationId": "get_player_profile_by_id_api_v1_players__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Stable player identifier.",
              "examples": [
                978838
              ],
              "title": "Player Id"
            },
            "description": "Stable player identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "The player identifier is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The player was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The player feature dataset is missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Player profile retrieval failed unexpectedly.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorCode": {
        "type": "string",
        "enum": [
          "player_not_found",
          "ambiguous_player",
          "dataset_unavailable",
          "invalid_dataset",
          "analysis_failed",
          "invalid_player_search",
          "player_search_failed",
          "invalid_player_profile",
          "player_profile_failed",
          "invalid_transfer_analysis_request"
        ]
      },
      "ApiErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail",
        "description": "Machine-readable and human-readable API error details."
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "description": "Standard error envelope returned by the API."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerInsightResponse": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "group_label": {
            "type": "string",
            "title": "Group Label"
          },
          "metric_key": {
            "type": "string",
            "title": "Metric Key"
          },
          "metric_label": {
            "type": "string",
            "title": "Metric Label"
          },
          "metric_short_label": {
            "type": "string",
            "title": "Metric Short Label"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "percentile": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percentile"
          },
          "peer_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Peer Count"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "group",
          "group_label",
          "metric_key",
          "metric_label",
          "metric_short_label",
          "value",
          "percentile",
          "peer_count",
          "evidence"
        ],
        "title": "PlayerInsightResponse",
        "description": "Explainable strength or watch-out."
      },
      "PlayerIntelligenceResponse": {
        "properties": {
          "position_group": {
            "type": "string",
            "title": "Position Group"
          },
          "sample": {
            "$ref": "#/components/schemas/PlayerSampleContextResponse"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/PlayerPerformanceMetricGroupResponse"
            },
            "type": "array",
            "title": "Groups"
          },
          "strengths": {
            "items": {
              "$ref": "#/components/schemas/PlayerInsightResponse"
            },
            "type": "array",
            "title": "Strengths"
          },
          "watch_outs": {
            "items": {
              "$ref": "#/components/schemas/PlayerInsightResponse"
            },
            "type": "array",
            "title": "Watch Outs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "position_group",
          "sample",
          "groups",
          "strengths",
          "watch_outs"
        ],
        "title": "PlayerIntelligenceResponse",
        "description": "Position-aware tournament intelligence."
      },
      "PlayerPerformanceMetricGroupResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/PlayerPerformanceMetricResponse"
            },
            "type": "array",
            "title": "Metrics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "metrics"
        ],
        "title": "PlayerPerformanceMetricGroupResponse",
        "description": "Grouped player metrics."
      },
      "PlayerPerformanceMetricResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "short_label": {
            "type": "string",
            "title": "Short Label"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "performance_percentile": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Performance Percentile"
          },
          "peer_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Peer Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "label",
          "short_label",
          "unit",
          "value",
          "peer_count"
        ],
        "title": "PlayerPerformanceMetricResponse",
        "description": "One position-aware tournament metric."
      },
      "PlayerProfileResponse": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "national_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "National Team Name"
          },
          "country_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "age": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "appearances": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appearances"
          },
          "starts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts"
          },
          "minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "weighted_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weighted Rating"
          },
          "market_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value"
          },
          "market_value_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value Currency"
          },
          "archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype"
          },
          "spatial_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spatial Role"
          },
          "final_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Role"
          },
          "lateral_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lateral Profile"
          },
          "vertical_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vertical Profile"
          },
          "mobility_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobility Profile"
          },
          "role_confidence_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Confidence Pct"
          },
          "spatial_reliability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spatial Reliability"
          },
          "data_reliability_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Reliability Score"
          },
          "player_quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Quality Score"
          },
          "role_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Reason"
          },
          "tournament": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerTournamentSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "intelligence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerIntelligenceResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "player_id",
          "player_name",
          "national_team_name",
          "country_name",
          "position",
          "age",
          "height_cm",
          "appearances",
          "starts",
          "minutes",
          "weighted_rating",
          "market_value",
          "market_value_currency",
          "archetype",
          "spatial_role",
          "final_role",
          "lateral_profile",
          "vertical_profile",
          "mobility_profile",
          "role_confidence_pct",
          "spatial_reliability",
          "data_reliability_score",
          "player_quality_score",
          "role_reason"
        ],
        "title": "PlayerProfileResponse",
        "description": "Detailed profile returned for one player."
      },
      "PlayerSampleContextResponse": {
        "properties": {
          "target_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Minutes"
          },
          "minimum_peer_minutes": {
            "type": "number",
            "minimum": 0.0,
            "title": "Minimum Peer Minutes"
          },
          "target_meets_peer_minimum": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Meets Peer Minimum"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_minutes",
          "minimum_peer_minutes",
          "target_meets_peer_minimum"
        ],
        "title": "PlayerSampleContextResponse",
        "description": "Sample-size context for percentile interpretation."
      },
      "PlayerSearchItemResponse": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "national_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "National Team Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "final_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Role"
          },
          "archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype"
          },
          "age": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "market_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value"
          },
          "market_value_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value Currency"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "player_id",
          "player_name",
          "national_team_name",
          "position",
          "final_role",
          "archetype",
          "age",
          "market_value",
          "market_value_currency"
        ],
        "title": "PlayerSearchItemResponse",
        "description": "Lightweight player information returned by search."
      },
      "PlayerSearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Count"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerSearchItemResponse"
            },
            "type": "array",
            "title": "Players"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "count",
          "players"
        ],
        "title": "PlayerSearchResponse",
        "description": "Structured response returned by player search."
      },
      "PlayerTournamentSummaryResponse": {
        "properties": {
          "matches": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matches"
          },
          "starts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts"
          },
          "substitute_appearances": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substitute Appearances"
          },
          "captain_appearances": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captain Appearances"
          },
          "minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "formations_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formations Used"
          },
          "primary_formation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Formation"
          },
          "primary_lineup_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Lineup Position"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "matches",
          "starts",
          "substitute_appearances",
          "captain_appearances",
          "minutes",
          "formations_used",
          "primary_formation",
          "primary_lineup_position"
        ],
        "title": "PlayerTournamentSummaryResponse",
        "description": "Tournament participation context for one player."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}